+++ src/pine/ast/where.clj
@@ -1,5 +1,7 @@
 (ns pine.ast.where)
 
 (defn handle [state [column operator value]]
-  (let [a (state :current)]
-    (update state :where conj [a column operator value])))
+  (let [a (state :current)
+        [alias col] (:value column)
+        alias (or alias a)]
+    (update state :where conj [alias col operator value])))